+++ src/test-functional/groovy/com/adarshr/gradle/testlogger/AbstractFunctionalSpec.groovy
@@ -15,6 +15,8 @@ abstract class AbstractFunctionalSpec extends Specification {
 
     private static final String TEST_ROOT = 'src/test-functional/resources'
 
+    private static final String GRADLE_VERSION = '4.7'
+
     @Rule
     TemporaryFolder temporaryFolder
 
@@ -64,7 +66,7 @@ abstract class AbstractFunctionalSpec extends Specification {
     private BuildResult runProject(File projectDir, String args) {
         try {
             GradleRunner.create()
-                .withGradleVersion('4.6')
+                .withGradleVersion(GRADLE_VERSION)
                 .withProjectDir(projectDir)
                 .withPluginClasspath()
                 .withDebug(true)