+++ src/test-functional/groovy/com/adarshr/gradle/testlogger/ThemeSwitchingSpec.groovy
@@ -17,10 +17,11 @@ class ThemeSwitchingSpec extends AbstractFunctionalSpec {
             def lines = getLoggerOutput(result.output).lines
         and:
             lines.size() == 4
-            lines[0] == render('com.adarshr.test.SingleSpec')
-            lines[1] == render('')
-            lines[2] == render('  Test this is a single test PASSED')
-            lines[3] == render('')
+            lines[0] == render('')
+            lines[1] == render('com.adarshr.test.SingleSpec')
+            lines[2] == render('')
+            lines[3] == render('  Test this is a single test PASSED')
+
         and:
             result.task(":test").outcome == SUCCESS
     }
@@ -36,10 +37,10 @@ class ThemeSwitchingSpec extends AbstractFunctionalSpec {
             def lines = getLoggerOutput(result.output).lines
         and:
             lines.size() == 4
-            lines[0] == render('com.adarshr.test.SingleSpec')
-            lines[1] == render('')
-            lines[2] == render('  Test this is a single test PASSED')
-            lines[3] == render('')
+            lines[0] == render('')
+            lines[1] == render('com.adarshr.test.SingleSpec')
+            lines[2] == render('')
+            lines[3] == render('  Test this is a single test PASSED')
         and:
             result.task(":test").outcome == SUCCESS
     }
@@ -55,10 +56,10 @@ class ThemeSwitchingSpec extends AbstractFunctionalSpec {
             def lines = getLoggerOutput(result.output).lines
         and:
             lines.size() == 4
-            lines[0] == render('[erase-ahead,bold,bright-yellow]com.adarshr.test.SingleSpec[/]')
-            lines[1] == render('')
-            lines[2] == render('[erase-ahead,bold]  Test [bold-off]this is a single test[green] PASSED[/]')
-            lines[3] == render('')
+            lines[0] == render('')
+            lines[1] == render('[erase-ahead,bold,bright-yellow]com.adarshr.test.SingleSpec[/]')
+            lines[2] == render('')
+            lines[3] == render('[erase-ahead,bold]  Test [bold-off]this is a single test[green] PASSED[/]')
         and:
             result.task(":test").outcome == SUCCESS
     }
@@ -74,12 +75,37 @@ class ThemeSwitchingSpec extends AbstractFunctionalSpec {
             def lines = getLoggerOutput(result.output).lines
         and:
             lines.size() == 4
-            lines[0] == render('  [erase-ahead,default]com.adarshr.test.SingleSpec[/]')
-            lines[1] == render('')
-            lines[2] == render("    [erase-ahead][green]${symbol}[grey] this is a single test[/]")
-            lines[3] == render('')
+            lines[0] == render('')
+            lines[1] == render('  [erase-ahead,default]com.adarshr.test.SingleSpec[/]')
+            lines[2] == render('')
+            lines[3] == render("    [erase-ahead][green]${symbol}[grey] this is a single test[/]")
         and:
             result.task(":test").outcome == SUCCESS
+    }
+
+    def "theme can be overriden using system property"() {
+        when:
+            def result = run(
+                'single-spock-test',
+                "testlogger { theme 'mocha' }",
+                'clean test -Dtestlogger.theme=plain'
+            )
+        then:
+            def lines = getLoggerOutput(result.output).lines
+        and:
+            lines.size() == 4
+            lines[0] == render('')
+            lines[1] == render('com.adarshr.test.SingleSpec')
+            lines[2] == render('')
+            lines[3] == render('  Test this is a single test PASSED')
+        and:
+            result.task(":test").outcome == SUCCESS
+        cleanup:
+            run(
+                'single-spock-test',
+                "testlogger { System.clearProperty('testlogger.theme') }",
+                'clean'
+            )
     }
 
     private static String getSymbol() {